Breaking up the blueprint file
