+++ mephisto/abstractions/_subcomponents/README.md
@@ -1,0 +1,3 @@
+# Abstraction Subcomponents
+
+This folder holds subcomponents of the core abstractions that are self-contained, but are built as a layer upon the data model. Each can be imported from its abstraction (like `from mephisto.abstractions.blueprint import TaskRunner`), but has logic better contained here.
