+++ mephisto/abstractions/architect.py
@@ -10,7 +10,7 @@ from omegaconf import MISSING, DictConfig
 from typing import Dict, List, Any, ClassVar, Type, TYPE_CHECKING, Callable
 
 if TYPE_CHECKING:
-    from mephisto.abstractions.channel import Channel
+    from mephisto.abstractions._subcomponents.channel import Channel
     from mephisto.data_model.packet import Packet
     from mephisto.data_model.task_run import TaskRun
     from mephisto.abstractions.database import MephistoDB
