+++ mephisto/abstractions/architects/ec2/ec2_architect.py
@@ -28,7 +28,7 @@ from mephisto.abstractions.architects.ec2.ec2_helpers import (
 )
 
 if TYPE_CHECKING:
-    from mephisto.abstractions.channel import Channel
+    from mephisto.abstractions._subcomponents.channel import Channel
     from mephisto.data_model.packet import Packet
     from mephisto.data_model.task_run import TaskRun
     from mephisto.abstractions.database import MephistoDB
